+++ apps/api/src/domain/power-manager/application/use-cases/delete-power-array.ts
@@ -1,18 +1,21 @@
 import { type Either, left, right } from '@/core/either';
+import { NotAllowedError } from '@/core/errors/not-allowed-error';
 import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error';
 import type { PowerArraysRepository } from '../repositories/power-arrays-repository';
 
 interface DeletePowerArrayUseCaseRequest {
   powerArrayId: string;
+  userId: string;
 }
 
-type DeletePowerArrayUseCaseResponse = Either<ResourceNotFoundError, Record<string, never>>;
+type DeletePowerArrayUseCaseResponse = Either<ResourceNotFoundError | NotAllowedError, null>;
 
 export class DeletePowerArrayUseCase {
   constructor(private powerArraysRepository: PowerArraysRepository) {}
 
   async execute({
     powerArrayId,
+    userId,
   }: DeletePowerArrayUseCaseRequest): Promise<DeletePowerArrayUseCaseResponse> {
     const powerArray = await this.powerArraysRepository.findById(powerArrayId);
 
@@ -20,8 +23,12 @@ export class DeletePowerArrayUseCase {
       return left(new ResourceNotFoundError());
     }
 
+    if (!powerArray.canBeEditedBy(userId)) {
+      return left(new NotAllowedError());
+    }
+
     await this.powerArraysRepository.delete(powerArrayId);
 
-    return right({});
+    return right(null);
   }
 }
